requirements.txt

streamlit
PyMuPDF
easyocr
Pillow
numpy
